requirements.txt
psutil==5.9.0
Pyrogram==1.4.4
requests==2.25.1
wget==3.2
tgcrypto
python-dotenv==0.19.2